+++ app/api/parcels.py
@@ -3,12 +3,10 @@ import datetime
 from flask import Flask, request, jsonify, make_response
 from flask_jwt_extended import JWTManager, create_access_token, get_jwt_identity, jwt_required
 from app.models.models import DatabaseConnection
+from app.api.helpers.validate_parcels import validate_parcel_info
 from app import app
 
-query = ""
-status = "Cancelled"
-
-db = None
+db = DatabaseConnection()
 
 @app.route('/api/v1/parcels', methods=['GET'])
 @jwt_required
@@ -19,7 +17,7 @@ def get_all_parcels():
 	current_user = get_jwt_identity()
 
 	query = """SELECT * FROM parcels;"""
-	connect_to_db()
+	db.connect()
 	db.cur.execute(query)
 	db.connection.commit()
 	result = db.cur.fetchall()
@@ -37,12 +35,9 @@ def get_parcel(parcel_id):
 		Function for API endpoint to fetch a specific parcel delivery order
 	"""
 	current_user = get_jwt_identity()
-
 	query = """SELECT * FROM parcels WHERE id = %s;"""
-	if type(parcel_id) != int:
-		return jsonify({'message': 'id should be an integer', 'status': 'failure'}), 400
 	
-	connect_to_db()
+	db.connect()
 	db.cur.execute(query, (parcel_id,))
 	db.connection.commit()
 	result = db.cur.fetchall()	
@@ -63,8 +58,8 @@ def cancel_order(parcel_id):
 	current_user = get_jwt_identity()
 	
 	query = """SELECT * FROM parcels WHERE id = %s AND status != %s;"""
-	connect_to_db()
-	db.cur.execute(query, (parcel_id, status,))
+	db.connect()
+	db.cur.execute(query, (parcel_id, 'Cancelled',))
 	result = db.cur.fetchall()
 	if result != None:
 		for row in result:
@@ -72,7 +67,7 @@ def cancel_order(parcel_id):
 				return jsonify({'message': 'parcel already delivered', 'status': 'failure'}), 400
 			else:
 				query = """UPDATE parcels SET status = %s WHERE id = %s"""
-				db.cur.execute(query, (status, parcel_id,))
+				db.cur.execute(query, ('Cancelled', parcel_id,))
 				db.connection.commit()
 				db.connection.close()
 				return jsonify({'message': 'parcel updated', 'status': 'success'}), 200			
@@ -101,7 +96,7 @@ def change_parcel_destination(parcel_id):
 			return jsonify({'message': 'destination should not be longer than 124 characters', 'status': 'failure'}), 400
 
 	query = """SELECT * FROM parcels WHERE id = %s;"""
-	connect_to_db()
+	db.connect()
 	db.cur.execute(query, (parcel_id,))
 	db.connection.commit()
 	result = db.cur.fetchall()
@@ -141,44 +136,13 @@ def create_order():
 	description = request.json['description'] 
 	if validate_parcel_info(owner, description, pickup_location, destination):
 		query = """INSERT INTO parcels (owner, description, date_created, pickup_location, present_location, destination, price, status) VALUES (%s, %s, %s, %s, %s, %s, %s, %s)"""
-		connect_to_db()
+		db.connect()
 		db.cur.execute(query, (owner, description, date_created, pickup_location, present_location, destination, price, status,))
 		db.connection.commit()
 		return jsonify({'message': 'parcel created', 'status': 'success'}), 200
 	else:
 		return jsonify({'message': 'parcel not created', 'status': 'failure'}), 400
 
-		#token here
 
-def connect_to_db():
-	global db
-	db = DatabaseConnection()
 	
 
-def validate_parcel_info(owner, description, pickup_location, destination):
-	"""
-		Function to validate parcel info. Making sure the required fields are filled in and correct
-	"""
-
-	if type(owner) != int: 
-		return jsonify({'message': 'parcel owner must be identified by id (integer)', 'status': 'failure'}), 400
-
-	if description == "": 
-		return jsonify({'message': 'description is empty', 'status': 'failure'}), 400
-	else:
-		if len(description) > 124:
-			return jsonify({'message': 'description should not be longer than 124 characters', 'status': 'failure'}), 400
-
-	if pickup_location == "": 
-		return jsonify({'message': 'pickup_location is empty', 'status': 'failure'}), 400
-	else:
-		if len(pickup_location) > 124:
-			return jsonify({'message': 'pickup location should not be longer than 124 characters', 'status': 'failure'}), 400
-
-	if destination == "": 
-		return jsonify({'message': 'destination is empty', 'status': 'failure'}), 400
-	else:
-		if len(destination) > 124:
-			return jsonify({'message': 'destination should not be longer than 124 characters', 'status': 'failure'}), 400	
-
-	return True
