+++ app/api/users.py
@@ -1,12 +1,13 @@
 import psycopg2
-import datetime, re
+import datetime
 from flask_jwt_extended import JWTManager, create_access_token, get_jwt_identity, jwt_required
 from app.models.models import DatabaseConnection
 from flask import Flask, request, jsonify, make_response
 from werkzeug.security import generate_password_hash, check_password_hash
+from app.api.helpers.validate_users import validate_user_info
 from app import app
 
-db = None
+db = DatabaseConnection()
 
 #api
 @app.route('/api/v1', methods=['GET'])
@@ -24,7 +25,7 @@ def get_all_users():
 	"""
 	current_user = get_jwt_identity()
 	query = """SELECT * FROM users;"""
-	connect_to_db()
+	db.connect()
 	db.cur.execute(query)
 	db.connection.commit()
 	result = db.cur.fetchall()
@@ -43,11 +44,9 @@ def get_user_parcel(user_id):
 		Function for API endpoint to fetch all parcel delivery orders by a specific user
 	"""
 	current_user = get_jwt_identity()
-	if type(user_id) != int:
-		return jsonify({'message':'user must be identified by an integer', 'status':'failure'}), 400
 
 	query = """SELECT * FROM parcels WHERE owner = %s;"""
-	connect_to_db()
+	db.connect()
 	db.cur.execute(query, (user_id,))
 	db.connection.commit()
 	result = db.cur.fetchall()
@@ -73,7 +72,7 @@ def login_user():
 	if validate_user_info(username, email, password, False):
 		password = generate_password_hash(password)
 		query = """SELECT * FROM users WHERE username = %s AND password_hash = %s;"""
-		connect_to_db()
+		db.connect()
 		db.cur.execute(query, (username, password,))
 		db.connection.commit()
 		result = db.cur.fetchall()
@@ -103,44 +102,14 @@ def create_user():
 
 	if validate_user_info(username, email, password, True):
 		query = """INSERT INTO users (username, email, password_hash) VALUES (%s, %s, %s)"""
-		connect_to_db()
+		db.connect()
 		db.cur.execute(query, (username, email, password,))
 		db.connection.commit()
 		db.cur.close()
 		db.connection.close()
-		return jsonify({'message': "user created"}), 201
+		access_token = create_access_token(identity = username)
+		return jsonify({'message': 'user signed up', 'status': 'success', 'access_token': access_token}), 201
 	else:
 		return jsonify({'message': "user not created", 'status': 'failure'}), 400
 
 
-def validate_user_info(username, email, password, signup):
-	"""
-		Function to validate user inputs. Making sure the required fields are filled in and correct
-	"""
-
-	if username == "": 
-		return jsonify({'message': 'username is empty', 'status': 'failure'}), 400
-	else:
-		if len(username) > 32:
-			return jsonify({'message': 'username should not longer than 32 characters', 'status': 'failure'}), 400	
-
-	if password == "": 
-		return jsonify({'message': 'password is empty', 'status': 'failure'}), 400
-	else:
-		if len(password) < 6:
-			return jsonify({'message': 'password should be longer than 6 characters', 'status': 'failure'}), 400	
-
-	if signup:
-		if email == "": 
-			return jsonify({'message': 'email address is empty', 'status': 'failure'}), 400
-		else:
-			pattern = r"(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)"
-			if not re.match(pattern, email):
-				return jsonify({'message': 'enter a valid email', 'status': 'failure'}), 400
-
-	return True
-
-
-def connect_to_db():
-	global db
-	db = DatabaseConnection()
